Make extension work for UITextView and UITextField
